requirements.txt
streamlit==1.50.0
google-genai==1.46.0
langchain-core==1.0.1
langchain-chroma==1.0.0
langchain_google_genai==3.0.0